switch long to intptr_t or int64_t, since long is 32-bit on 64-bit windows

+++ cpp/cpp_odbc/Library/cpp_odbc/environment.h
@@ -32,18 +32,18 @@ public:
 	std::shared_ptr<connection> make_connection(std::string const & connection_string) const;
 
 	/**
-	 * @param Set the attribute to the given long value
+	 * @param Set the attribute to the given intptr_t value
 	 * @param attribute An ODBC constant which represents the attribute which shall be set
 	 * @param value The new value for the attribute
 	 */
-	void set_attribute(SQLINTEGER attribute, long value) const;
+	void set_attribute(SQLINTEGER attribute, intptr_t value) const;
 
 	virtual ~environment();
 protected:
 	environment();
 private:
 	virtual std::shared_ptr<connection> do_make_connection(std::string const & connection_string) const = 0;
-	virtual void do_set_attribute(SQLINTEGER attribute, long value) const = 0;
+	virtual void do_set_attribute(SQLINTEGER attribute, intptr_t value) const = 0;
 };
 
 }
